add check that startdate is before enddate.
