+++ Src/MoneyFox.Domain/Exceptions/StartAfterEnddateException.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Runtime.Serialization;
+
+namespace MoneyFox.Domain.Exceptions
+{
+    [Serializable]
+    public class StartAfterEnddateException : Exception
+    {
+        public StartAfterEnddateException()
+        {
+        }
+
+        protected StartAfterEnddateException(SerializationInfo info, StreamingContext context) : base(info, context)
+        {
+        }
+
+        public StartAfterEnddateException(string message) : base(message)
+        {
+        }
+
+        public StartAfterEnddateException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+    }
+}
